extract examples from test files
