+++ ratelimiter/ratelimiter_e2e_test.go
@@ -5,7 +5,6 @@ package ratelimiter_test
 
 import (
 	"context"
-	"fmt"
 	"os"
 	"testing"
 	"time"
@@ -178,75 +177,5 @@ func TestRatelimiter(t *testing.T) {
 	)
 	if err == nil {
 		t.Fatal("Resource must not be acquired")
-	}
-}
-
-func Example() {
-	ctx := context.TODO()
-	db, err := ydb.Open(ctx, "grpcs://localhost:2135/?database=/local")
-	if err != nil {
-		fmt.Printf("failed to connect: %v", err)
-		return
-	}
-	defer db.Close(ctx) // cleanup resources
-	// create node
-	err = db.Coordination().CreateNode(ctx, testCoordinationNodePath, coordination.NodeConfig{
-		Path:                     "",
-		SelfCheckPeriodMillis:    1000,
-		SessionGracePeriodMillis: 1000,
-		ReadConsistencyMode:      coordination.ConsistencyModeRelaxed,
-		AttachConsistencyMode:    coordination.ConsistencyModeRelaxed,
-		RatelimiterCountersMode:  coordination.RatelimiterCountersModeDetailed,
-	})
-	if err != nil {
-		fmt.Printf("failed to create node: %v", err)
-		return
-	}
-	defer func() {
-		// cleanup node
-		err = db.Coordination().DropNode(ctx, testCoordinationNodePath)
-		if err != nil {
-			fmt.Printf("failed to drop node: %v", err)
-		}
-	}()
-	// create resource
-	err = db.Ratelimiter().CreateResource(ctx, testCoordinationNodePath, ratelimiter.Resource{
-		ResourcePath: testResource,
-		HierarchicalDrr: ratelimiter.HierarchicalDrrSettings{
-			MaxUnitsPerSecond:       1,
-			MaxBurstSizeCoefficient: 2,
-		},
-	})
-	if err != nil {
-		fmt.Printf("failed to create resource: %v", err)
-	}
-	defer func() {
-		// cleanup resource
-		err = db.Ratelimiter().DropResource(ctx, testCoordinationNodePath, testResource)
-		if err != nil {
-			fmt.Printf("failed to drop resource: %v", err)
-		}
-	}()
-	// alter resource
-	err = db.Ratelimiter().AlterResource(ctx, testCoordinationNodePath, ratelimiter.Resource{
-		ResourcePath: testResource,
-		HierarchicalDrr: ratelimiter.HierarchicalDrrSettings{
-			MaxUnitsPerSecond:       3,
-			MaxBurstSizeCoefficient: 4,
-		},
-	})
-	if err != nil {
-		fmt.Printf("failed to alter resource: %v", err)
-	}
-	// acquire resource amount 1
-	err = db.Ratelimiter().AcquireResource(
-		ctx,
-		testCoordinationNodePath,
-		testResource,
-		1,
-		ratelimiter.WithAcquire(),
-	)
-	if err != nil {
-		fmt.Printf("failed to acquire resource: %v", err)
 	}
 }
